Refactor typeCheckExpression to Expression#typecheck
